- rename class from UrlPanel to ButtonUrlPanel - change button layout to fill empty space - use log instead system print for link opening errors
